requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
redis==5.0.1
pymongo==4.6.0

# HTTP requests and web scraping
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# AI and ML
openai==1.3.7
google-generativeai==0.3.2
pytrends==4.9.2

# Data processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Task scheduling
apscheduler==3.10.4
celery==5.3.4

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 